+++ tasks/move-granules/index.js
@@ -5,28 +5,32 @@ const { DuplicateFile, InvalidArgument } = require('@cumulus/common/errors');
 const get = require('lodash.get');
 const clonedeep = require('lodash.clonedeep');
 const flatten = require('lodash.flatten');
+const path = require('path');
+
 const {
   getRenamedS3File, unversionFilename,
   moveGranuleFile, renameS3FileWithTimestamp
 } = require('@cumulus/ingest/granule');
+
 const {
-  getCmrXMLFiles, getGranuleId
+  getCmrXMLFiles,
+  getGranuleId,
+  isECHO10File,
+  metadataObjectFromCMRXMLFile,
+  updateEcho10XMLMetadata
 } = require('@cumulus/cmrjs');
-const urljoin = require('url-join');
-const path = require('path');
+
 const {
   aws: {
     buildS3Uri,
     checksumS3Objects,
     deleteS3Object,
     parseS3Uri,
-    promiseS3Upload,
-    s3ObjectExists,
-    s3TagSetToQueryString
-  }
+    s3ObjectExists
+  },
+  BucketsConfig
 } = require('@cumulus/common');
 const { urlPathTemplate } = require('@cumulus/ingest/url-path-template');
-const xml2js = require('xml2js');
 const log = require('@cumulus/common/log');
 
 
@@ -85,7 +89,7 @@ function addInputFilesToGranules(input, granules, regex) {
  * validates the file matched only one collection.file and has a valid bucket
  * config.
  * @param {Array<Object>} match - list of matched collection.file
- * @param {Object} buckets - the buckets involved with the files
+ * @param {BucketsConfig} buckets - instance describing stack configuration.
  * @param {Object} file - the fileObject tested.
  * @throws {InvalidArgument} - If match is invalid, throws and error.
  */
@@ -97,7 +101,7 @@ function validateMatch(match, buckets, file) {
     throw new InvalidArgument(`File (${file}) did not match any collection.regexp.`);
   }
 
-  if (!buckets[match[0].bucket]) {
+  if (!buckets.keyExists(match[0].bucket)) {
     throw new InvalidArgument(`Collection config specifies a bucket key of ${match[0].bucket}, `
                               + `but the configured bucket keys are: ${Object.keys(buckets).join(', ')}`);
   }
@@ -110,46 +114,47 @@ function validateMatch(match, buckets, file) {
 * @param {Object} granulesObject - an object of granules where the key is the granuleId
 * @param {Object} collection - configuration object defining a collection
 *                              of granules and their files
-* @param {string} cmrFiles - array of objects that include CMR xmls uris and granuleIds
-* @param {Object} buckets - the buckets involved with the files
+* @param {Array<Object>} cmrFiles - array of objects that include CMR xmls uris and granuleIds
+* @param {BucketsConfig} buckets -  instance associated with the stack
 * @returns {Object} new granulesObject where each granules' files are updated with
 *                   the correct target buckets/paths/and s3uri filenames.
 **/
-function updateGranuleMetadata(granulesObject, collection, cmrFiles, buckets) {
+async function updateGranuleMetadata(granulesObject, collection, cmrFiles, buckets) {
   const updatedGranules = {};
   const fileSpecs = collection.files;
-  Object.keys(granulesObject).forEach((granuleId) => {
+
+  await Promise.all(Object.keys(granulesObject).map(async (granuleId) => {
     const updatedFiles = [];
     updatedGranules[granuleId] = { ...granulesObject[granuleId] };
 
-    granulesObject[granuleId].files.forEach((file) => {
+    const cmrFile = cmrFiles.find((f) => f.granuleId === granuleId);
+    const cmrMetadata = cmrFile ? await metadataObjectFromCMRXMLFile(cmrFile.filename) : {};
+
+    await Promise.all(granulesObject[granuleId].files.map(async (file) => {
       const match = fileSpecs.filter((cf) => unversionFilename(file.name).match(cf.regex));
       validateMatch(match, buckets, file);
 
       const URLPathTemplate = file.url_path || match[0].url_path || collection.url_path || '';
-      const cmrFile = cmrFiles.find((f) => f.granuleId === granuleId);
       const urlPath = urlPathTemplate(URLPathTemplate, {
         file,
         granule: granulesObject[granuleId],
-        cmrMetadata: cmrFile ? cmrFile.metadataObject : {}
+        cmrMetadata
       });
-      const bucket = buckets[match[0].bucket];
+      const bucketName = buckets.nameByKey(match[0].bucket);
       const filepath = path.join(urlPath, file.name);
 
       updatedFiles.push({
         ...file,
         ...{
-          bucket,
+          bucket: bucketName,
           filepath,
-          filename: `s3://${path.join(bucket.name, filepath)}`,
+          filename: `s3://${path.join(bucketName, filepath)}`,
           url_path: URLPathTemplate
         }
       });
-    });
-
+    }));
     updatedGranules[granuleId].files = [...updatedFiles];
-  });
-
+  }));
   return updatedGranules;
 }
 
@@ -160,9 +165,10 @@ function updateGranuleMetadata(granulesObject, collection, cmrFiles, buckets) {
  * @param {Object} file - granule file to be moved
  * @param {string} sourceBucket - source bucket location of files
  * @param {string} duplicateHandling - how to handle duplicate files
+ * @param {BucketsConfig} buckets - BucketsConfig instance
  * @returns {Array<Object>} returns the file moved and the renamed existing duplicates if any
  */
-async function moveFileRequest(file, sourceBucket, duplicateHandling) {
+async function moveFileRequest(file, sourceBucket, duplicateHandling, buckets) {
   const fileStagingDir = file.fileStagingDir || 'file-staging';
   const source = {
     Bucket: sourceBucket,
@@ -170,7 +176,7 @@ async function moveFileRequest(file, sourceBucket, duplicateHandling) {
   };
 
   const target = {
-    Bucket: file.bucket.name,
+    Bucket: file.bucket,
     Key: file.filepath
   };
 
@@ -191,7 +197,7 @@ async function moveFileRequest(file, sourceBucket, duplicateHandling) {
 
   if (s3ObjAlreadyExists && duplicateHandling === 'skip') return [fileMoved];
 
-  const options = (file.bucket.type.match('public')) ? { ACL: 'public-read' } : null;
+  const options = (buckets.type(file.bucket).match('public')) ? { ACL: 'public-read' } : null;
 
   // compare the checksum of the existing file and new file, and handle them accordingly
   if (s3ObjAlreadyExists && duplicateHandling === 'version') {
@@ -218,7 +224,7 @@ async function moveFileRequest(file, sourceBucket, duplicateHandling) {
   // return both file moved and renamed files
   return [fileMoved]
     .concat(renamedFiles.map((f) => ({
-      bucket: file.bucket,
+      bucket: f.Bucket,
       name: path.basename(f.Key),
       filename: buildS3Uri(f.Bucket, f.Key),
       filepath: f.Key,
@@ -234,18 +240,19 @@ async function moveFileRequest(file, sourceBucket, duplicateHandling) {
 * @param {Object} granulesObject - an object of the granules where the key is the granuleId
 * @param {string} sourceBucket - source bucket location of files
 * @param {string} duplicateHandling - how to handle duplicate files
+* @param {BucketsConfig} buckets - BucketsConfig instance
 * @returns {Object} the object with updated granules
 **/
-async function moveFilesForAllGranules(granulesObject, sourceBucket, duplicateHandling) {
+async function moveFilesForAllGranules(granulesObject, sourceBucket, duplicateHandling, buckets) {
   const moveFileRequests = Object.keys(granulesObject).map(async (granuleKey) => {
     const granule = granulesObject[granuleKey];
     const cmrFileFormat = /.*\.cmr\.xml$/;
     const filesToMove = granule.files.filter((file) => !file.name.match(cmrFileFormat));
-    const cmrFile = granule.files.filter((file) => file.name.match(cmrFileFormat));
-
-    const filesMoved = await Promise.all(filesToMove.map((file) =>
-      moveFileRequest(file, sourceBucket, duplicateHandling)));
-    granule.files = flatten(filesMoved).concat(cmrFile);
+    const cmrFiles = granule.files.filter((file) => file.name.match(cmrFileFormat));
+    const filesMoved = await Promise.all(
+      filesToMove.map((file) => moveFileRequest(file, sourceBucket, duplicateHandling, buckets))
+    );
+    granule.files = flatten(filesMoved).concat(flatten(cmrFiles));
   });
 
   await Promise.all(moveFileRequests);
@@ -253,69 +260,34 @@ async function moveFilesForAllGranules(granulesObject, sourceBucket, duplicateHa
 }
 
 /**
-* Update the online access url fields in CMR xml files
+ * Updates a CMR file modifying/adding an OnlineResources/OnlineAccessURLs
+ * element to the metadata for the new files locations.
+ *
+ * @param {Object} cmrFile - CMR file object (from getCmrXMLFiles)
+ * @param {Array<Object>} files - List of granule file objects for the CMR object
+ * @param {string} distEndpoint - distribution endpoint
+ * @param {BucketsConfig} buckets - Stack BucketsConfig instance
+ */
+async function updateCMRFileAccessURLs(cmrFile, files, distEndpoint, buckets) {
+  const updatedCmrFile = files.find((f) => isECHO10File(f.filename));
+  await moveGranuleFile(parseS3Uri(cmrFile.filename), parseS3Uri(updatedCmrFile.filename));
+  return updateEcho10XMLMetadata(updatedCmrFile, files, distEndpoint, buckets);
+}
+
+/**
+* Update each of the CMR files' onlineaccessurl fields to represent the new
+* file locations.
 *
-* @param {string} cmrFiles - array of objects that include CMR xmls uris and granuleIds
+* @param {Array<Object>} cmrFiles - array of objects that include CMR xmls uris and granuleIds
 * @param {Object} granulesObject - an object of the granules where the key is the granuleId
 * @param {string} distEndpoint - the api distribution endpoint
+* @param {BucketsConfig} buckets - BucketsConfig instance
 * @returns {Promise} promise resolves when all files have been updated
 **/
-async function updateCmrFileAccessURLs(cmrFiles, granulesObject, distEndpoint) {
-  await Promise.all(cmrFiles.map(async (cmrFile) => {
-    const metadataGranule = get(cmrFile, 'metadataObject.Granule');
+async function updateEachCmrFileAccessURLs(cmrFiles, granulesObject, distEndpoint, buckets) {
+  return Promise.all(cmrFiles.map(async (cmrFile) => {
     const granule = granulesObject[cmrFile.granuleId];
-    const urls = [];
-    // Populates onlineAcessUrls with all public and protected files
-    granule.files.forEach((file) => {
-      const urlObj = {};
-      if (file.bucket.type.match('protected')) {
-        const extension = urljoin(file.bucket.name, file.filepath);
-        urlObj.URL = urljoin(distEndpoint, extension);
-        urlObj.URLDescription = 'File to download';
-        urls.push(urlObj);
-        log.info(`protected file: ${JSON.stringify(file)},\nurl: ${JSON.stringify(urlObj)}`);
-      }
-      else if (file.bucket.type.match('public')) {
-        urlObj.URL = `https://${file.bucket.name}.s3.amazonaws.com/${file.filepath}`;
-        urlObj.URLDescription = 'File to download';
-        urls.push(urlObj);
-      }
-    });
-
-    const updatedGranule = {};
-    Object.keys(metadataGranule).forEach((key) => {
-      if (key === 'OnlineResources' || key === 'Orderable') {
-        updatedGranule.OnlineAccessURLs = {};
-      }
-      updatedGranule[key] = metadataGranule[key];
-    });
-    updatedGranule.OnlineAccessURLs.OnlineAccessURL = urls;
-    /* eslint-disable no-param-reassign */
-    cmrFile.metadataObject.Granule = updatedGranule;
-
-    const builder = new xml2js.Builder();
-    const xml = builder.buildObject(cmrFile.metadataObject);
-    cmrFile.metadata = xml;
-    /* eslint-enable no-param-reassign */
-    const updatedCmrFile = granule.files.find((f) => f.filename.match(/.*\.cmr\.xml$/));
-    // S3 upload only accepts tag query strings, so reduce tags to query string.
-    const tagsQueryString = s3TagSetToQueryString(cmrFile.s3Tags);
-    const params = {
-      Bucket: updatedCmrFile.bucket.name,
-      Key: updatedCmrFile.filepath,
-      Body: xml,
-      Tagging: tagsQueryString
-    };
-    if (updatedCmrFile.bucket.type.match('public')) {
-      params.ACL = 'public-read';
-      await promiseS3Upload(params);
-    }
-    else {
-      await promiseS3Upload(params);
-    }
-    // clean up old CmrFile after uploading new one
-    const { Bucket, Key } = parseS3Uri(cmrFile.filename);
-    await deleteS3Object(Bucket, Key);
+    return updateCMRFileAccessURLs(cmrFile, granule.files, distEndpoint, buckets);
   }));
 }
 
@@ -367,7 +339,7 @@ async function moveGranules(event) {
   // first we check if there is an output file
   const config = get(event, 'config');
   const bucket = get(config, 'bucket'); // the name of the bucket with private/public keys
-  const buckets = get(config, 'buckets');
+  const buckets = new BucketsConfig(get(config, 'buckets'));
   const regex = get(config, 'granuleIdExtraction', '(.*)');
   const inputGranules = get(config, 'input_granules', {});
   const distEndpoint = get(config, 'distribution_endpoint');
@@ -376,38 +348,34 @@ async function moveGranules(event) {
 
   const duplicateHandling = duplicateHandlingType(event);
 
-  const input = get(event, 'input', []);
+  const inputFileList = get(event, 'input', []);
 
   // Get list of cmr file objects from the input Array of S3 filenames (in
   // staging location after processing)
-  const cmrFiles = await getCmrXMLFiles(input, regex);
+  const cmrFiles = getCmrXMLFiles(inputFileList, regex);
 
   // create granules object for cumulus indexer
-  const allGranules = addInputFilesToGranules(input, inputGranules, regex);
+  const allGranules = addInputFilesToGranules(inputFileList, inputGranules, regex);
 
-  const granulesToMove = updateGranuleMetadata(allGranules, collection, cmrFiles, buckets);
-
-  // allows us to disable moving the files
+  let granulesToMove;
   let movedGranules;
+  // allows us to disable moving the files
   if (moveStagedFiles) {
+    granulesToMove = await updateGranuleMetadata(allGranules, collection, cmrFiles, buckets);
+
     // move files from staging location to final location
-    movedGranules = await moveFilesForAllGranules(granulesToMove, bucket, duplicateHandling);
+    movedGranules = await moveFilesForAllGranules(
+      granulesToMove, bucket, duplicateHandling, buckets
+    );
     // update cmr.xml files with correct online access urls
-    await updateCmrFileAccessURLs(cmrFiles, movedGranules, distEndpoint);
+    await updateEachCmrFileAccessURLs(cmrFiles, movedGranules, distEndpoint, buckets);
+  }
+  else {
+    movedGranules = allGranules;
   }
 
   return {
-    granules: Object.keys(movedGranules).map((k) => {
-      const granule = movedGranules[k];
-
-      // Just return the bucket name with the granules
-      granule.files.map((f) => {
-        f.bucket = f.bucket.name; /* eslint-disable-line no-param-reassign */
-        return f;
-      });
-
-      return granule;
-    })
+    granules: Object.keys(movedGranules).map((k) => movedGranules[k])
   };
 }